fix some compiler warnings
git-svn-id: svn://svn.code.sf.net/p/tntdb/code/trunk@332 8aebba32-5d12-0410-a889-bbfed4e15866

+++ tntdb/test/base-test.cpp
@@ -302,9 +302,9 @@ class TntdbBaseTest : public cxxtools::unit::TestSuite
         CXXTOOLS_UNIT_ASSERT(shortNotNull);
         CXXTOOLS_UNIT_ASSERT(longNotNull);
         CXXTOOLS_UNIT_ASSERT(!ushortNotNull);
-        CXXTOOLS_UNIT_ASSERT_EQUALS(intVal, rowcount + 3);
-        CXXTOOLS_UNIT_ASSERT_EQUALS(shortVal, rowcount + 4);
-        CXXTOOLS_UNIT_ASSERT_EQUALS(longVal, rowcount + 5);
+        CXXTOOLS_UNIT_ASSERT_EQUALS(static_cast<unsigned>(intVal), rowcount + 3);
+        CXXTOOLS_UNIT_ASSERT_EQUALS(static_cast<unsigned>(shortVal), rowcount + 4);
+        CXXTOOLS_UNIT_ASSERT_EQUALS(static_cast<unsigned>(longVal), rowcount + 5);
       }
     }
 
@@ -370,9 +370,9 @@ class TntdbBaseTest : public cxxtools::unit::TestSuite
         CXXTOOLS_UNIT_ASSERT(shortNotNull);
         CXXTOOLS_UNIT_ASSERT(longNotNull);
         CXXTOOLS_UNIT_ASSERT(!ushortNotNull);
-        CXXTOOLS_UNIT_ASSERT_EQUALS(intVal, rowcount + 4);
-        CXXTOOLS_UNIT_ASSERT_EQUALS(shortVal, rowcount + 5);
-        CXXTOOLS_UNIT_ASSERT_EQUALS(longVal, rowcount + 6);
+        CXXTOOLS_UNIT_ASSERT_EQUALS(static_cast<unsigned>(intVal), rowcount + 4);
+        CXXTOOLS_UNIT_ASSERT_EQUALS(static_cast<unsigned>(shortVal), rowcount + 5);
+        CXXTOOLS_UNIT_ASSERT_EQUALS(static_cast<unsigned>(longVal), rowcount + 6);
       }
     }
